+++ src/main/java/org/sa46lll/client/OutputView.java
@@ -8,17 +8,17 @@ public class OutputView {
     private static final String WELCOME_MESSAGE = "에 오신 것을 환영합니다.";
     private static final String SELECTED_INFO_MESSAGE = "%s는 %d원 입니다.";
 
-    public static void printWelcomeMessage(String store) {
+    static void printWelcomeMessage(String store) {
         System.out.println(store + WELCOME_MESSAGE);
     }
 
-    public static void printProducts(List<Product> products) {
+    static void printProducts(List<Product> products) {
         products.forEach(
                 product -> System.out.println("- " + product.getName())
         );
     }
 
-    public static void printSelectedInfo(Product product) {
+    static void printSelectedInfo(Product product) {
         System.out.printf(SELECTED_INFO_MESSAGE, product.getName(), product.getPrice());
     }
 }
